+++ src/components/SideBar/SidebarFooter.tsx
@@ -3,93 +3,87 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 const styles = {
-  footerContainer: {
-    width: "100%",
-  },
   footerItem: {
     textDecoration: "none",
-    paddingRight: "0.75rem",
+    paddingRight: 10,
   },
   footerItemText: {
     "&:hover": {
       textDecoration: "underline",
     },
     color: "gray.main",
-    fontSize: "0.75rem",
+    fontSize: 12,
   },
   footerNav: {
     display: "flex",
     flexWrap: "wrap",
-    height: "100%",
     paddingX: 2,
   },
 };
 
 const SidebarFooter = () => {
   return (
-    <Box sx={styles.footerContainer}>
-      <Box sx={styles.footerNav}>
-        <Link
-          target="_blank"
-          to="/coming-soon"
-          rel="noreferrer"
-          style={styles.footerItem}
-        >
-          <Typography variant="subtitle2" sx={styles.footerItemText}>
-            Terms of Service
-          </Typography>
-        </Link>
-        <Link
-          target="_blank"
-          to="/coming-soon"
-          rel="noreferrer"
-          style={styles.footerItem}
-        >
-          <Typography variant="subtitle2" sx={styles.footerItemText}>
-            Privacy Policy
-          </Typography>
-        </Link>
-        <Link
-          target="_blank"
-          to="/coming-soon"
-          rel="noreferrer"
-          style={styles.footerItem}
-        >
-          <Typography variant="subtitle2" sx={styles.footerItemText}>
-            Contact Us
-          </Typography>
-        </Link>
-        <Link
-          target="_blank"
-          to="/coming-soon"
-          rel="noreferrer"
-          style={styles.footerItem}
-        >
-          <Typography variant="subtitle2" sx={styles.footerItemText}>
-            About The Team
-          </Typography>
-        </Link>
-        <Link
-          target="_blank"
-          to="https://github.com/orgs/Project-Chirp/repositories"
-          rel="noreferrer"
-          style={styles.footerItem}
-        >
-          <Typography variant="subtitle2" sx={styles.footerItemText}>
-            GitHub Repo
-          </Typography>
-        </Link>
-        <Link
-          target="_blank"
-          to="/coming-soon"
-          rel="noreferrer"
-          style={styles.footerItem}
-        >
-          <Typography variant="subtitle2" sx={styles.footerItemText}>
-            Docs
-          </Typography>
-        </Link>
-      </Box>
+    <Box component={"nav"} sx={styles.footerNav}>
+      <Link
+        target="_blank"
+        to="/coming-soon"
+        rel="noreferrer"
+        style={styles.footerItem}
+      >
+        <Typography variant="subtitle2" sx={styles.footerItemText}>
+          Terms of Service
+        </Typography>
+      </Link>
+      <Link
+        target="_blank"
+        to="/coming-soon"
+        rel="noreferrer"
+        style={styles.footerItem}
+      >
+        <Typography variant="subtitle2" sx={styles.footerItemText}>
+          Privacy Policy
+        </Typography>
+      </Link>
+      <Link
+        target="_blank"
+        to="/coming-soon"
+        rel="noreferrer"
+        style={styles.footerItem}
+      >
+        <Typography variant="subtitle2" sx={styles.footerItemText}>
+          Contact Us
+        </Typography>
+      </Link>
+      <Link
+        target="_blank"
+        to="/coming-soon"
+        rel="noreferrer"
+        style={styles.footerItem}
+      >
+        <Typography variant="subtitle2" sx={styles.footerItemText}>
+          About The Team
+        </Typography>
+      </Link>
+      <Link
+        target="_blank"
+        to="https://github.com/orgs/Project-Chirp/repositories"
+        rel="noreferrer"
+        style={styles.footerItem}
+      >
+        <Typography variant="subtitle2" sx={styles.footerItemText}>
+          GitHub Repo
+        </Typography>
+      </Link>
+      <Link
+        target="_blank"
+        to="/coming-soon"
+        rel="noreferrer"
+        style={styles.footerItem}
+      >
+        <Typography variant="subtitle2" sx={styles.footerItemText}>
+          Docs
+        </Typography>
+      </Link>
     </Box>
   );
 };
